+++ src/main/java/org/jabref/gui/groups/GroupTreeView.java
@@ -613,4 +613,11 @@ public class GroupTreeView extends BorderPane {
             }
         }
     }
+
+    /**
+     * Focus on GroupTree
+     */
+    public void requestFocusGroupTree() {
+        groupTree.requestFocus();
+    }
 }